+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 3.3.0 - 2025-09-20
-*   **Fonctionnalité :** Ajout d'un système d'envoi automatique d'emails d'anniversaire aux adhérents.
+*   **Fonctionnalité :** Ajout d'un système d'envoi automatique d'emails d'anniversaire aux adhérents, avec une option pour l'activer ou le désactiver.
 *   **Fonctionnalité :** Ajout d'une option dans `Réglages > Options DAME` pour définir le slug d'un article (publié ou privé) à utiliser comme modèle pour l'email d'anniversaire.
 *   **Fonctionnalité :** L'email d'anniversaire est personnalisé avec les balises `[NOM]`, `[PRENOM]` et `[AGE]`.
 *   **Fonctionnalité :** Une nouvelle tâche cron journalière envoie les emails d'anniversaire. Un email de résumé est envoyé à l'administrateur.